+++ SYMPHONY/src/LP/lp_rowfunc.c
@@ -384,10 +384,11 @@ void add_waiting_rows(lp_prob *p, waiting_row **wrows, int add_row_num)
 
    sense = lp_data->tmp.c; /* m */
    rhs = lp_data->tmp.d; /* m */
-   REMALLOC(lp_data->tmp.dv, double, lp_data->tmp.dv_size, nzcnt, 5*BB_BUNCH);
+   REMALLOC(lp_data->tmp.dv, double, lp_data->tmp.dv_size, nzcnt, 
+         5*(int)BB_BUNCH);
    rmatval = lp_data->tmp.dv; /* nzcnt */
    rmatbeg = lp_data->tmp.i1;
-   REMALLOC(lp_data->tmp.iv, int, lp_data->tmp.iv_size, nzcnt, 5*BB_BUNCH);
+   REMALLOC(lp_data->tmp.iv, int, lp_data->tmp.iv_size, nzcnt, 5*(int)BB_BUNCH);
    rmatind = lp_data->tmp.iv;
 
    *rmatbeg = 0;
